fix routes class unit test
