+++ packages/stacked_generator/lib/src/generators/router/generator/routes_class_builder.dart
@@ -1,0 +1,96 @@
+import 'package:code_builder/code_builder.dart';
+import 'package:stacked_generator/route_config_resolver.dart';
+
+class RoutesClassBuilder {
+  final List<RouteConfig> routes;
+
+  const RoutesClassBuilder({
+    required this.routes,
+  });
+
+  /// Example result
+  ///
+  /// class RoutesClassName {
+  /// static const String loginView = 'pathNamaw';
+  /// static const String _homeView = '/family/:fid';
+  /// static String homeView({@required dynamic fid}) => '/family/\$fid';
+  /// static const all = <String>{
+  /// loginView,_homeView,};}
+  ///
+  Class buildRoutesClass() {
+    final assignPathsToRouteNames = routes.map((route) => Field(
+          (b) => b
+            ..modifier = FieldModifier.constant
+            ..static = true
+            ..name = _convertToPrivateNameWhenRouteHasPathParameter(route)
+            ..assignment = literalString(route.pathName).code,
+        ));
+
+    final pathMethods =
+        routes.where((route) => route.pathName.contains(':')).map((route) {
+      return Method(
+        (b) => b
+          ..returns = Reference('String')
+          ..static = true
+          ..lambda = true
+          ..optionalParameters
+              .addAll(_extractOptionalParameters(route.pathName))
+          ..name = route.name
+          ..body = Code(_addRouteWithPathParameter(
+              routeName: route.name, routePath: route.pathName)),
+      );
+    });
+
+    final allField = Field(
+      (b) => b
+        ..modifier = FieldModifier.constant
+        ..static = true
+        ..name = 'all'
+        ..assignment = literalSet(
+                routes.map((route) => Reference(
+                    _convertToPrivateNameWhenRouteHasPathParameter(route))),
+                Reference('String'))
+            .code,
+    );
+
+    return Class((b) => b
+      ..name = 'Routes'
+      ..methods.addAll(pathMethods)
+      ..fields.addAll([...assignPathsToRouteNames, allField]));
+  }
+
+  String _convertToPrivateNameWhenRouteHasPathParameter(RouteConfig route) {
+    return route.pathName.contains(':') ? '_${route.name}' : route.name;
+  }
+
+  Iterable<Parameter> _extractOptionalParameters(String pathName) {
+    return RegExp(r':([^/]+)').allMatches(pathName).map((m) {
+      final match = m.group(1);
+      if (match!.endsWith('?')) {
+        return Parameter((b) => b
+          ..named = true
+          ..type = Reference('dynamic')
+          ..name = match.substring(0, match.length - 1));
+      } else {
+        return Parameter((b) => b
+          ..required = true
+          ..named = true
+          ..type = Reference('dynamic')
+          ..name = match);
+      }
+    });
+  }
+
+  String _addRouteWithPathParameter(
+      {required String routePath, required String routeName}) {
+    final pathAfterAddingDollarSigns =
+        routePath.replaceAllMapped(RegExp(r'([:])|([?])'), (m) {
+      if (m[1] != null) {
+        return '\$';
+      } else {
+        return '';
+      }
+    });
+    return "'${pathAfterAddingDollarSigns}'";
+  }
+}
